refactor to clean code on project validations
