--- src/client/src/utils/validations/index.ts
@@ -1,5 +1,0 @@
-import * as projectName from "./projectName";
-
-export {
-  projectName
-};
